Refactor codebase for better organization and maintainability
Refactor the codebase to separate client and server functionality into distinct modules and improve maintainability.

* **Client Code (`client.c`):**
  - Include `client_operations.h` for client-specific operations.
  - Remove OpenSSL library includes and common functions.
  - Update `main` function to use new client-specific functions for upload, download, and delete operations.

* **Server Code (`server.c`):**
  - Include `server_operations.h` for server-specific operations.
  - Remove OpenSSL library includes and common functions.
  - Update `main` function to use new server-specific functions for handling write, read, and delete requests.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/gurelbs/TFTP?shareId=XXXX-XXXX-XXXX-XXXX).
