+++ common.h
@@ -11,5 +11,7 @@ void calculate_md5(FILE *file, unsigned char *digest);
 int encrypt_data(unsigned char *plaintext, int plaintext_len, unsigned char *ciphertext);
 int decrypt_data(unsigned char *ciphertext, int ciphertext_len, unsigned char *plaintext);
 void init_aes_keys();
+void set_socket_timeout(int socket, int sec, int usec);
+void ensure_backup_dir(const char *backup_dir);
 
 #endif
